src/backend/requirements.txt
# Base packages
python-dotenv
fastapi
uvicorn[standard]
python-multipart
pydantic[email]

# Azure SDK Core
azure-core
requests
aiohttp

# Azure Services
azure-identity
azure-storage-blob
azure-cosmos
azure-servicebus
azure-mgmt-storage
azure-functions

# Development tools
pytest
pytest-mock
black
pylint
flake8
flake8-docstrings
flake8-import-order
flake8-bugbear
flake8-annotations

# Additional utilities
aiofiles
structlog
typing-extensions
python-jose[cryptography]
passlib[bcrypt]
semantic-kernel[azure]==1.32.2
openai==2.0.0
sqlparse
sqlglot
unittest2
json_utils
starlette
aiortc
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-otlp-proto-http
azure-monitor-events-extension
opentelemetry-sdk==1.37.0
opentelemetry-api==1.37.0
opentelemetry-semantic-conventions==0.58b0
opentelemetry-instrumentation==0.58b0
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-requests
azure-monitor-opentelemetry==1.8.1
azure-ai-projects==1.0.0
